# BACK-END/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff E-Commerce ",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "This is the Login Page for Admin or User",
        "produces": [
          "application/json"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login for Admin or User based on role",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login for Admin or User based on role",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/signup": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "This is the Signup Page for User",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Signup for new User",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Signup for Admin or User based on role",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSignup"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route produces render adminPanel.ejs",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Get Brands Page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/categories": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Get Categories Page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/allusers": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route gets all users by Admin. This route produces render adminPanel.ejs user routes for json response",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route gets all users by Admin. This route produces render adminPanel.ejs user routes for json response",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/login": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "This is the Login Page for only admin",
        "produces": [
          "application/json"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Login for Admin for front page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login for Admin for front page",
            "required": true,
            "schema": {
              "$ref": "#/definitions/frontadminlogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/registeredUsers/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "This route render page for registered users. This route produces render userPanel.ejs",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Login": {
      "type": "object",
      "properties": {
        "identifier": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      },
      "required": [
        "identifier",
        "password"
      ]
    },
    "UserSignup": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string",
          "example": "John"
        },
        "lastname": {
          "type": "string",
          "example": "Doe"
        },
        "username": {
          "type": "string",
          "example": "johndoe"
        },
        "email": {
          "type": "string",
          "example": "johndoe@example.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        },
        "address": {
          "type": "string",
          "example": "Oslo"
        },
        "telephonenumber": {
          "type": "string",
          "example": "12345678"
        }
      },
      "required": [
        "firstname",
        "lastname",
        "username",
        "email",
        "password",
        "address",
        "telephonenumber"
      ]
    },
...
    "frontadminlogin": {
      "type": "object",
      "properties": {
        "identifier": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      },
      "required": [
        "identifier",
        "password"
      ]
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}